+++ Events.php
@@ -11,7 +11,7 @@ namespace humhub\modules\sharebetween;
 use humhub\modules\activity\models\Activity;
 use humhub\modules\content\components\ContentActiveRecord;
 use humhub\modules\content\events\ContentEvent;
-use humhub\modules\sharebetween\activities\SharedContentCreated;
+use humhub\modules\notification\models\Notification;
 use humhub\modules\sharebetween\models\Share;
 use yii\base\BaseObject;
 
@@ -62,7 +62,18 @@ class Events extends BaseObject
 
         if ($activity->object_model === Share::class) {
             // Switch to render specific text for the shared content activity
-            $activity->class = SharedContentCreated::class;
+            $activity->class = activities\SharedContentCreated::class;
+        }
+    }
+
+    public static function onNotificationAfterFind($event)
+    {
+        /* @var Notification $activity */
+        $notification = $event->sender;
+
+        if ($notification->source_class === Share::class) {
+            // Switch to render specific text for the shared content notification
+            $notification->class = notifications\SharedContentCreated::class;
         }
     }
 }