+++ docs/CHANGELOG.md
@@ -3,7 +3,7 @@ Changelog
 
 1.0.8 (Unreleased)
 -------------------------
-- Enh #67: Update text in the activity sidebar
+- Enh #67: Update text for activity and notification of shared contents
 
 1.0.7 (November 14, 2024)
 -------------------------